backend/requirements.txt
connexion[swagger-ui,flask,uvicorn] >3.0.0;
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
werkzeug == 3.0.1;
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
# should support both Python 2 and Python 3
flask-cors >= 3.0.10
setuptools >= 21.0.0
Flask == 3.0.2
flask-migrate == 4.0.5
flask-sqlalchemy == 3.1.1
openai == 1.12.0
python-dotenv == 1.0.1
psycopg2 == 2.9.9
